+++ src/Pyro4/core.py
@@ -42,7 +42,6 @@ class URI(object):
     automatically be converted to uppercase internally.
     """
     uriRegEx = re.compile(r"(?P<protocol>[Pp][Yy][Rr][Oo][a-zA-Z]*):(?P<object>\S+?)(@(?P<location>\S+))?$")
-    __slots__ = ("protocol", "object", "sockname", "host", "port", "object")
 
     def __init__(self, uri):
         if isinstance(uri, URI):